+++ RELEASES.md
@@ -38,6 +38,7 @@ This will be the final release with support for the dYdX v3 (legacy) API. Future
 - Added Polymarket data loader event-level API support (#3484), thanks @jsemldonado
 - Added Polymarket `event_slug_builder` support (#3501), thanks @jsemldonado
 - Added Polymarket batch order support (#3506), thanks @loafer-19
+- Added Tardis data client with factory in Rust
 - Improved tearsheet with dynamic Nautilus version and refined run info table (#3396), thanks @KaulSe
 
 ### Breaking Changes